Move profile classes to profiles module
