+++ aws-auth/src/main/java/software/amazon/awssdk/awsauth/regions/internal/PartitionMetadataProvider.java
@@ -20,12 +20,12 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import software.amazon.awssdk.annotations.SdkInternalApi;
-import software.amazon.awssdk.awsauth.regions.PartitionServiceMetadata;
 import software.amazon.awssdk.awsauth.regions.Region;
 import software.amazon.awssdk.awsauth.regions.RegionMetadata;
 import software.amazon.awssdk.awsauth.regions.ServiceMetadata;
 import software.amazon.awssdk.awsauth.regions.ServiceMetadataProvider;
 import software.amazon.awssdk.awsauth.regions.internal.model.Partition;
+import software.amazon.awssdk.awsauth.regions.providers.PartitionServiceMetadata;
 import software.amazon.awssdk.utils.Validate;
 
 /**
