+++ aws-auth/src/test/java/software/amazon/awssdk/awsauth/regions/providers/InstanceProfileRegionProviderTest.java
@@ -93,6 +93,4 @@ public class InstanceProfileRegionProviderTest {
         }
 
     }
-
-
 }
