+++ aws-core/src/main/java/software/amazon/awssdk/awscore/endpoint/EndpointUtils.java
@@ -13,12 +13,11 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.awscore.util;
+package software.amazon.awssdk.awscore.endpoint;
 
 import java.net.URI;
 import software.amazon.awssdk.annotations.SdkInternalApi;
 import software.amazon.awssdk.awsauth.regions.Region;
-import software.amazon.awssdk.awsauth.util.DefaultServiceEndpointBuilder;
 
 /**
  * A collection of utility methods centered around generating service endpoints from various pieces of information.
