+++ pom.xml
@@ -58,6 +58,7 @@
         <module>aws-auth</module>
         <module>aws-core</module>
         <module>core</module>
+        <module>profiles</module>
         <module>bom</module>
         <module>codegen</module>
         <module>http-client-spi</module>
