+++ profiles/src/test/java/software/amazon/awssdk/profiles/ProfileFileTest.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.awsauth.credentials.profile;
+package software.amazon.awssdk.profiles;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -21,11 +21,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.AbstractMap;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.function.Consumer;
 import java.util.stream.Stream;
 import org.junit.Test;
-import software.amazon.awssdk.awsauth.credentials.AwsSessionCredentials;
-import software.amazon.awssdk.awsauth.regions.Region;
 import software.amazon.awssdk.core.util.StringInputStream;
 
 /**
@@ -40,59 +37,59 @@ public class ProfileFileTest {
     @Test
     public void emptyProfilesHaveNoProperties() {
         assertThat(configFileProfiles("[profile foo]"))
-                .isEqualTo(profiles(profile("foo")));
+            .isEqualTo(profiles(profile("foo")));
     }
 
     @Test
     public void profileDefinitionsMustEndWithBrackets() {
         assertThatThrownBy(() -> configFileProfiles("[profile foo"))
-                .hasMessageContaining("Profile definition must end with ']'");
+            .hasMessageContaining("Profile definition must end with ']'");
     }
 
     @Test
     public void profileNamesShouldBeTrimmed() {
         assertThat(configFileProfiles("[profile \tfoo \t]"))
-                .isEqualTo(profiles(profile("foo")));
+            .isEqualTo(profiles(profile("foo")));
     }
 
     @Test
     public void tabsCanSeparateProfileNamesFromProfilePrefix() {
         assertThat(configFileProfiles("[profile\tfoo]"))
-                .isEqualTo(profiles(profile("foo")));
+            .isEqualTo(profiles(profile("foo")));
     }
 
     @Test
     public void propertiesMustBeDefinedInAProfile() {
         assertThatThrownBy(() -> configFileProfiles("name = value"))
-                .hasMessageContaining("Expected a profile definition");
+            .hasMessageContaining("Expected a profile definition");
     }
 
     @Test
     public void profilesCanContainProperties() {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "name = value"))
-                .isEqualTo(profiles(profile("foo", property("name", "value"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value"))));
     }
 
     @Test
     public void windowsStyleLineEndingsAreSupported() {
         assertThat(configFileProfiles("[profile foo]\r\n" +
                                       "name = value"))
-                .isEqualTo(profiles(profile("foo", property("name", "value"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value"))));
     }
 
     @Test
     public void equalsSignsAreSupportedInPropertyNames() {
         assertThat(configFileProfiles("[profile foo]\r\n" +
                                       "name = val=ue"))
-                .isEqualTo(profiles(profile("foo", property("name", "val=ue"))));
+            .isEqualTo(profiles(profile("foo", property("name", "val=ue"))));
     }
 
     @Test
     public void unicodeCharactersAreSupportedInPropertyValues() {
         assertThat(configFileProfiles("[profile foo]\r\n" +
                                       "name = \uD83D\uDE02"))
-                .isEqualTo(profiles(profile("foo", property("name", "\uD83D\uDE02"))));
+            .isEqualTo(profiles(profile("foo", property("name", "\uD83D\uDE02"))));
     }
 
     @Test
@@ -100,45 +97,45 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "name = value\n" +
                                       "name2 = value2"))
-                .isEqualTo(profiles(profile("foo",
-                                            property("name", "value"),
-                                            property("name2", "value2"))));
+            .isEqualTo(profiles(profile("foo",
+                                        property("name", "value"),
+                                        property("name2", "value2"))));
     }
 
     @Test
     public void propertyKeysAndValuesAreTrimmed() {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "name \t=  \tvalue \t"))
-                .isEqualTo(profiles(profile("foo", property("name", "value"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value"))));
     }
 
     @Test
     public void propertyValuesCanBeEmpty() {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "name ="))
-                .isEqualTo(profiles(profile("foo", property("name", ""))));
+            .isEqualTo(profiles(profile("foo", property("name", ""))));
     }
 
     @Test
     public void propertyKeysCannotBeEmpty() {
         assertThatThrownBy(() -> configFileProfiles("[profile foo]\n" +
                                                     "= value"))
-                .hasMessageContaining("Property did not have a name");
+            .hasMessageContaining("Property did not have a name");
     }
 
     @Test
     public void propertyDefinitionsMustContainAnEqualsSign() {
         assertThatThrownBy(() -> configFileProfiles("[profile foo]\n" +
                                                     "key : value"))
-                .hasMessageContaining("Expected an '=' sign defining a property");
+            .hasMessageContaining("Expected an '=' sign defining a property");
     }
 
     @Test
     public void multipleProfilesCanBeEmpty() {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "[profile bar]"))
-                .isEqualTo(profiles(profile("foo"),
-                                    profile("bar")));
+            .isEqualTo(profiles(profile("foo"),
+                                profile("bar")));
     }
 
     @Test
@@ -147,8 +144,8 @@ public class ProfileFileTest {
                                       "name = value\n" +
                                       "[profile bar]\n" +
                                       "name2 = value2"))
-                .isEqualTo(profiles(profile("foo", property("name", "value")),
-                                    profile("bar", property("name2", "value2"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value")),
+                                profile("bar", property("name2", "value2"))));
     }
 
     @Test
@@ -161,8 +158,8 @@ public class ProfileFileTest {
                                       "\t \n" +
                                       "[profile bar]\n" +
                                       " \t"))
-                .isEqualTo(profiles(profile("foo", property("name", "value")),
-                                    profile("bar")));
+            .isEqualTo(profiles(profile("foo", property("name", "value")),
+                                profile("bar")));
     }
 
     @Test
@@ -170,7 +167,7 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("# Comment\n" +
                                       "[profile foo] # Comment\n" +
                                       "name = value # Comment with # sign"))
-                .isEqualTo(profiles(profile("foo", property("name", "value"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value"))));
     }
 
     @Test
@@ -178,7 +175,7 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("; Comment\n" +
                                       "[profile foo] ; Comment\n" +
                                       "name = value ; Comment with ; sign"))
-                .isEqualTo(profiles(profile("foo", property("name", "value"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value"))));
     }
 
     @Test
@@ -186,7 +183,7 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("# Comment\n" +
                                       "[profile foo] ; Comment\n" +
                                       "name = value # Comment with ; sign"))
-                .isEqualTo(profiles(profile("foo", property("name", "value"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value"))));
     }
 
     @Test
@@ -194,15 +191,15 @@ public class ProfileFileTest {
         assertThat(configFileProfiles(";\n" +
                                       "[profile foo];\n" +
                                       "name = value ;\n"))
-                .isEqualTo(profiles(profile("foo", property("name", "value"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value"))));
     }
 
     @Test
     public void commentsCanBeAdjacentToProfileNames() {
         assertThat(configFileProfiles("[profile foo]; Adjacent semicolons\n" +
                                       "[profile bar]# Adjacent pound signs"))
-                .isEqualTo(profiles(profile("foo"),
-                                    profile("bar")));
+            .isEqualTo(profiles(profile("foo"),
+                                profile("bar")));
     }
 
     @Test
@@ -210,9 +207,9 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "name = value; Adjacent semicolons\n" +
                                       "name2 = value# Adjacent pound signs"))
-                .isEqualTo(profiles(profile("foo",
-                                            property("name", "value; Adjacent semicolons"),
-                                            property("name2", "value# Adjacent pound signs"))));
+            .isEqualTo(profiles(profile("foo",
+                                        property("name", "value; Adjacent semicolons"),
+                                        property("name2", "value# Adjacent pound signs"))));
     }
 
     @Test
@@ -220,8 +217,8 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "name = value\n" +
                                       " -continued"))
-                .isEqualTo(profiles(profile("foo",
-                                            property("name", "value\n-continued"))));
+            .isEqualTo(profiles(profile("foo",
+                                        property("name", "value\n-continued"))));
     }
 
     @Test
@@ -230,8 +227,8 @@ public class ProfileFileTest {
                                       "name = value\n" +
                                       " -continued\n" +
                                       " -and-continued"))
-                .isEqualTo(profiles(profile("foo",
-                                            property("name", "value\n-continued\n-and-continued"))));
+            .isEqualTo(profiles(profile("foo",
+                                        property("name", "value\n-continued\n-and-continued"))));
     }
 
     @Test
@@ -239,21 +236,21 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "name = value\n" +
                                       " -continued ; Comment"))
-                .isEqualTo(profiles(profile("foo",
-                                            property("name", "value\n-continued ; Comment"))));
+            .isEqualTo(profiles(profile("foo",
+                                        property("name", "value\n-continued ; Comment"))));
     }
 
     @Test
     public void continuationsCannotBeUsedOutsideOfAProfile() {
         assertThatThrownBy(() -> configFileProfiles(" -continued"))
-                .hasMessageContaining("Expected a profile or property definition");
+            .hasMessageContaining("Expected a profile or property definition");
     }
 
     @Test
     public void continuationsCannotBeUsedOutsideOfAProperty() {
         assertThatThrownBy(() -> configFileProfiles("[profile foo]\n" +
                                                     " -continued"))
-                .hasMessageContaining("Expected a profile or property definition");
+            .hasMessageContaining("Expected a profile or property definition");
     }
 
     @Test
@@ -262,7 +259,7 @@ public class ProfileFileTest {
                                                     "name = value\n" +
                                                     "[profile foo]\n" +
                                                     " -continued"))
-                .hasMessageContaining("Expected a profile or property definition");
+            .hasMessageContaining("Expected a profile or property definition");
     }
 
     @Test
@@ -271,9 +268,9 @@ public class ProfileFileTest {
                                       "name = value\n" +
                                       "[profile foo]\n" +
                                       "name2 = value2"))
-                .isEqualTo(profiles(profile("foo",
-                                            property("name", "value"),
-                                            property("name2", "value2"))));
+            .isEqualTo(profiles(profile("foo",
+                                        property("name", "value"),
+                                        property("name2", "value2"))));
     }
 
     @Test
@@ -281,7 +278,7 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "name = value\n" +
                                       "name = value2"))
-                .isEqualTo(profiles(profile("foo", property("name", "value2"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value2"))));
     }
 
     @Test
@@ -290,7 +287,7 @@ public class ProfileFileTest {
                                       "name = value\n" +
                                       "[profile foo]\n" +
                                       "name = value2"))
-                .isEqualTo(profiles(profile("foo", property("name", "value2"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value2"))));
     }
 
     @Test
@@ -299,7 +296,7 @@ public class ProfileFileTest {
                                       "name = value\n" +
                                       "[default]\n" +
                                       "name2 = value2"))
-                .isEqualTo(profiles(profile("default", property("name", "value"))));
+            .isEqualTo(profiles(profile("default", property("name", "value"))));
     }
 
     @Test
@@ -308,7 +305,7 @@ public class ProfileFileTest {
                                       "name2 = value2\n" +
                                       "[profile default]\n" +
                                       "name = value"))
-                .isEqualTo(profiles(profile("default", property("name", "value"))));
+            .isEqualTo(profiles(profile("default", property("name", "value"))));
     }
 
     @Test
@@ -317,26 +314,27 @@ public class ProfileFileTest {
                                          "name = value\n",
                                          "[in valid 2]\n" +
                                          "name2 = value2"))
-                .isEqualTo(profiles());
+            .isEqualTo(profiles());
     }
 
     @Test
     public void invalidPropertyNamesAreIgnored() {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "in valid = value"))
-                .isEqualTo(profiles(profile("foo")));
+            .isEqualTo(profiles(profile("foo")));
     }
 
     @Test
     public void allValidProfileNameCharactersAreSupported() {
         assertThat(configFileProfiles("[profile ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_]"))
-                .isEqualTo(profiles(profile("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_")));
+            .isEqualTo(profiles(profile("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_")));
     }
 
     @Test
     public void allValidPropertyNameCharactersAreSupported() {
         assertThat(configFileProfiles("[profile foo]\nABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_ = value"))
-                .isEqualTo(profiles(profile("foo", property("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_", "value"))));
+            .isEqualTo(profiles(profile("foo", property("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_",
+                                                        "value"))));
     }
 
     @Test
@@ -344,7 +342,7 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "s3 =\n" +
                                       " name = value"))
-                .isEqualTo(profiles(profile("foo", property("s3", "\nname = value"))));
+            .isEqualTo(profiles(profile("foo", property("s3", "\nname = value"))));
     }
 
     @Test
@@ -352,7 +350,7 @@ public class ProfileFileTest {
         assertThatThrownBy(() -> configFileProfiles("[profile foo]\n" +
                                                     "s3 =\n" +
                                                     " invalid"))
-                .hasMessageContaining("Expected an '=' sign defining a property");
+            .hasMessageContaining("Expected an '=' sign defining a property");
     }
 
     @Test
@@ -360,7 +358,7 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "s3 =\n" +
                                       " name ="))
-                .isEqualTo(profiles(profile("foo", property("s3", "\nname ="))));
+            .isEqualTo(profiles(profile("foo", property("s3", "\nname ="))));
     }
 
     @Test
@@ -368,7 +366,7 @@ public class ProfileFileTest {
         assertThatThrownBy(() -> configFileProfiles("[profile foo]\n" +
                                                     "s3 =\n" +
                                                     " = value"))
-                .hasMessageContaining("Property did not have a name");
+            .hasMessageContaining("Property did not have a name");
     }
 
     @Test
@@ -376,7 +374,7 @@ public class ProfileFileTest {
         assertThat(configFileProfiles("[profile foo]\n" +
                                       "s3 =\n" +
                                       " in valid = value"))
-                .isEqualTo(profiles(profile("foo", property("s3", "\nin valid = value"))));
+            .isEqualTo(profiles(profile("foo", property("s3", "\nin valid = value"))));
     }
 
     @Test
@@ -386,7 +384,7 @@ public class ProfileFileTest {
                                       " name = value\n" +
                                       "\t \n" +
                                       " name2 = value2"))
-                .isEqualTo(profiles(profile("foo", property("s3", "\nname = value\nname2 = value2"))));
+            .isEqualTo(profiles(profile("foo", property("s3", "\nname = value\nname2 = value2"))));
     }
 
     @Test
@@ -395,9 +393,9 @@ public class ProfileFileTest {
                                          "name = value\n",
                                          "[foo]\n" +
                                          "name2 = value2"))
-                .isEqualTo(profiles(profile("foo",
-                                            property("name", "value"),
-                                            property("name2", "value2"))));
+            .isEqualTo(profiles(profile("foo",
+                                        property("name", "value"),
+                                        property("name2", "value2"))));
     }
 
     @Test
@@ -408,9 +406,9 @@ public class ProfileFileTest {
                                       "name2 = value2\n" +
                                       "[profile default]\n" +
                                       "name3 = value3"))
-                .isEqualTo(profiles(profile("default",
-                                            property("name", "value"),
-                                            property("name3", "value3"))));
+            .isEqualTo(profiles(profile("default",
+                                        property("name", "value"),
+                                        property("name3", "value3"))));
     }
 
     @Test
@@ -419,156 +417,26 @@ public class ProfileFileTest {
                                          "name = value",
                                          "[foo]\n" +
                                          "name = value2"))
-                .isEqualTo(profiles(profile("foo", property("name", "value2"))));
+            .isEqualTo(profiles(profile("foo", property("name", "value2"))));
     }
 
     @Test
     public void configProfilesWithoutPrefixAreIgnored() {
         assertThat(configFileProfiles("[foo]\n" +
                                       "name = value"))
-                .isEqualTo(profiles());
+            .isEqualTo(profiles());
     }
 
     @Test
     public void credentialsProfilesWithPrefixAreIgnored() {
         assertThat(credentialFileProfiles("[profile foo]\n" +
                                           "name = value"))
-                .isEqualTo(profiles());
-    }
-
-    @Test
-    public void roleProfileCanInheritFromAnotherFile() {
-        String sourceProperties =
-                "aws_access_key_id=defaultAccessKey\n" +
-                "aws_secret_access_key=defaultSecretAccessKey";
-
-        String childProperties =
-                "source_profile=source\n" +
-                "role_arn=arn:aws:iam::123456789012:role/testRole";
-
-        String configSource = "[profile source]\n" + sourceProperties;
-        String credentialsSource = "[source]\n" + sourceProperties;
-        String configChild = "[profile child]\n" + childProperties;
-        String credentialsChild = "[child]\n" + childProperties;
-
-        Consumer<Map<String, Profile>> profileValidator = p ->
-            assertThatThrownBy(() -> p.get("child").credentialsProvider())
-                    .hasMessageContaining("the 'sts' service module must be on the class path");
-
-        assertThat(aggregateFileProfiles(configSource, credentialsChild)).satisfies(profileValidator);
-        assertThat(aggregateFileProfiles(configChild, credentialsSource)).satisfies(profileValidator);
-    }
-
-    @Test
-    public void roleProfileWithMissingSourceThrowsException() {
-        assertThatThrownBy(() -> configFile("[profile test]\n" +
-                                            "source_profile=source\n" +
-                                            "role_arn=arn:aws:iam::123456789012:role/testRole").profile("test").get()
-                                                                                               .credentialsProvider())
-                .hasMessageContaining("source profile has no credentials configured.");
-    }
-
-    @Test
-    public void roleProfileWithSourceThatHasNoCredentialsThrowsExceptionWhenLoadingCredentials() {
-        ProfileFile profiles = configFile("[profile source]\n" +
-                                          "[profile test]\n" +
-                                          "source_profile=source\n" +
-                                          "role_arn=arn:aws:iam::123456789012:role/testRole");
-
-        assertThat(profiles.profile("test")).hasValueSatisfying(profile -> {
-            assertThatThrownBy(profile::credentialsProvider).hasMessageContaining("source profile has no credentials configured");
-        });
-    }
-
-    @Test
-    public void profileFileWithRegionLoadsCorrectly() {
-        assertThat(allTypesProfile().profile("profile-with-region")).hasValueSatisfying(profile -> {
-            assertThat(profile.region()).hasValue(Region.of("us-east-1"));
-        });
-    }
-
-    @Test
-    public void profileFileWithStaticCredentialsLoadsCorrectly() {
-        assertThat(allTypesProfile().profile("default")).hasValueSatisfying(profile -> {
-            assertThat(profile.name()).isEqualTo("default");
-            assertThat(profile.property(ProfileProperties.AWS_ACCESS_KEY_ID)).hasValue("defaultAccessKey");
-            assertThat(profile.toString()).contains("default");
-            assertThat(profile.region()).isNotPresent();
-            assertThat(profile.credentialsProvider()).hasValueSatisfying(credentialsProvider -> {
-                assertThat(credentialsProvider.getCredentials()).satisfies(credentials -> {
-                    assertThat(credentials.accessKeyId()).isEqualTo("defaultAccessKey");
-                    assertThat(credentials.secretAccessKey()).isEqualTo("defaultSecretAccessKey");
-                });
-            });
-        });
-    }
-
-    @Test
-    public void profileFileWithSessionCredentialsLoadsCorrectly() {
-        assertThat(allTypesProfile().profile("profile-with-session-token")).hasValueSatisfying(profile -> {
-            assertThat(profile.region()).isNotPresent();
-            assertThat(profile.credentialsProvider()).hasValueSatisfying(credentialsProvider -> {
-                assertThat(credentialsProvider.getCredentials()).satisfies(credentials -> {
-                    assertThat(credentials).isInstanceOf(AwsSessionCredentials.class);
-                    assertThat(credentials.accessKeyId()).isEqualTo("defaultAccessKey");
-                    assertThat(credentials.secretAccessKey()).isEqualTo("defaultSecretAccessKey");
-                    assertThat(((AwsSessionCredentials) credentials).sessionToken()).isEqualTo("awsSessionToken");
-                });
-            });
-        });
-    }
-
-    @Test
-    public void profileFileWithAssumeRoleThrowsExceptionWhenRetrievingCredentialsProvider() {
-        assertThat(allTypesProfile().profile("profile-with-assume-role")).hasValueSatisfying(profile -> {
-            assertThat(profile.region()).isNotPresent();
-            assertThatThrownBy(profile::credentialsProvider).isInstanceOf(IllegalStateException.class);
-        });
-    }
-
-    @Test
-    public void profileFileWithCircularDependencyThrowsExceptionWhenResolvingCredentials() {
-        assertThatThrownBy(() -> configFile("[profile source]\n" +
-                                            "aws_access_key_id=defaultAccessKey\n" +
-                                            "aws_secret_access_key=defaultSecretAccessKey\n" +
-                                            "\n" +
-                                            "[profile test]\n" +
-                                            "source_profile=test3\n" +
-                                            "role_arn=arn:aws:iam::123456789012:role/testRole\n" +
-                                            "\n" +
-                                            "[profile test2]\n" +
-                                            "source_profile=test\n" +
-                                            "role_arn=arn:aws:iam::123456789012:role/testRole2\n" +
-                                            "\n" +
-                                            "[profile test3]\n" +
-                                            "source_profile=test2\n" +
-                                            "role_arn=arn:aws:iam::123456789012:role/testRole3").profile("test").get()
-                                                                                                .credentialsProvider())
-                .isInstanceOf(IllegalStateException.class)
-                .hasMessageContaining("Invalid profile file: Circular relationship detected with profiles");
+            .isEqualTo(profiles());
     }
 
     @Test
     public void loadingDefaultProfileFileWorks() {
         ProfileFile.defaultProfileFile();
-    }
-
-    private ProfileFile allTypesProfile() {
-        return configFile("[default]\n" +
-                          "aws_access_key_id = defaultAccessKey\n" +
-                          "aws_secret_access_key = defaultSecretAccessKey\n" +
-                          "\n" +
-                          "[profile profile-with-session-token]\n" +
-                          "aws_access_key_id = defaultAccessKey\n" +
-                          "aws_secret_access_key = defaultSecretAccessKey\n" +
-                          "aws_session_token = awsSessionToken\n" +
-                          "\n" +
-                          "[profile profile-with-region]\n" +
-                          "region = us-east-1\n" +
-                          "\n" +
-                          "[profile profile-with-assume-role]\n" +
-                          "source_profile=default\n" +
-                          "role_arn=arn:aws:iam::123456789012:role/testRole\n");
     }
 
     private ProfileFile configFile(String configFile) {
