+++ services/sts/src/main/java/software/amazon/awssdk/services/sts/internal/StsProfileCredentialsProviderFactory.java
@@ -18,12 +18,12 @@ package software.amazon.awssdk.services.sts.internal;
 import software.amazon.awssdk.annotations.SdkProtectedApi;
 import software.amazon.awssdk.awsauth.credentials.AwsCredentials;
 import software.amazon.awssdk.awsauth.credentials.AwsCredentialsProvider;
-import software.amazon.awssdk.awsauth.credentials.profile.Profile;
-import software.amazon.awssdk.awsauth.credentials.profile.ProfileProperties;
-import software.amazon.awssdk.awsauth.credentials.profile.internal.ChildProfileCredentialsProviderFactory;
+import software.amazon.awssdk.awsauth.credentials.ChildProfileCredentialsProviderFactory;
 import software.amazon.awssdk.awsauth.regions.Region;
 import software.amazon.awssdk.awsauth.regions.providers.AwsRegionProviderChain;
 import software.amazon.awssdk.awsauth.regions.providers.DefaultAwsRegionProviderChain;
+import software.amazon.awssdk.profiles.Profile;
+import software.amazon.awssdk.profiles.ProfileProperties;
 import software.amazon.awssdk.services.sts.STSClient;
 import software.amazon.awssdk.services.sts.auth.StsAssumeRoleCredentialsProvider;
 import software.amazon.awssdk.services.sts.model.AssumeRoleRequest;
