+++ services/sts/src/test/java/software/amazon/awssdk/services/sts/internal/AssumeRoleProfileTest.java
@@ -18,8 +18,9 @@ package software.amazon.awssdk.services.sts.internal;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.Test;
-import software.amazon.awssdk.awsauth.credentials.profile.ProfileFile;
+import software.amazon.awssdk.awsauth.credentials.internal.ProfileCredentialsUtils;
 import software.amazon.awssdk.core.util.StringInputStream;
+import software.amazon.awssdk.profiles.ProfileFile;
 import software.amazon.awssdk.services.sts.AssumeRoleIntegrationTest;
 import software.amazon.awssdk.utils.SdkAutoCloseable;
 
@@ -44,7 +45,7 @@ public class AssumeRoleProfileTest {
                                           .type(ProfileFile.Type.CONFIGURATION)
                                           .build();
         assertThat(profiles.profile("test")).hasValueSatisfying(profile -> {
-            assertThat(profile.credentialsProvider()).hasValueSatisfying(credentialsProvider -> {
+            assertThat(new ProfileCredentialsUtils(profile, profiles::profile).credentialsProvider()).hasValueSatisfying(credentialsProvider -> {
                 assertThat(credentialsProvider).isInstanceOf(SdkAutoCloseable.class);
                 ((SdkAutoCloseable) credentialsProvider).close();
             });
